Music and Buttons
All 3 buttons now working as well as music has been added. Imported time so that when you enter the rules menu it gives you 2s to read then displays the back button. The music can be changed by changing the name of the file in the game. also rules need to be added